Restructure the preprocessor files
The restructuring brings some clarity into the preprocessor library
files. Specifically we now have two subdirectories:
 - utils, with utility classes that are more general than the
   preprocessor.
 - mesh, the set of files needed to implement the mesh, but not those
   that use the mesh.
