+++ applications/Preprocessor/extractSurface.cpp
@@ -86,7 +86,8 @@ using idx_t = std::size_t;
 #include <string>
 #include "Base/CommandLineOptions.h"
 #include "Base/MpiContainer.h"
-#include "mesh.h"
+#include "mesh/Mesh.h"
+#include "MeshFactory.h"
 #include "hpgem.h"
 #include "centaur.h"
 #include "meshData.h"
