+++ applications/Preprocessor/lib/output_impl.h
@@ -41,7 +41,7 @@
 #include <fstream>
 #include <set>
 #include <map>
-#include "tag.h"
+#include "utils/tag.h"
 
 namespace Preprocessor {
