+++ applications/Preprocessor/main.cpp
@@ -48,7 +48,7 @@ using idx_t = std::size_t;
 #include <random>
 #include "Base/CommandLineOptions.h"
 #include "Base/MpiContainer.h"
-#include "mesh.h"
+#include "mesh/Mesh.h"
 #include "MeshFactory.h"
 #include "hpgem.h"
 #include "centaur.h"
